+++ src/Symfony/Component/Messenger/Bridge/Amqp/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 5.2.0
 -----
 
+ * Introduced support for multiple bindings on the same queue.
  * Add option to confirm message delivery
  * DSN now support AMQPS out-of-the-box.